# swagger/swagger.yaml
definitions:
  jsonobject.BatchItem:
    properties:
      correlation_id:
        example: "1"
        type: string
      original_url:
        description: URL для сокращения
        example: http://ya.ru
        type: string
      short_url:
        description: Сокращенный URL
        example: http://localhost:8080/rjhsha
        type: string
    type: object
  jsonobject.Response:
    properties:
      result:
        example: http://localhost:8080/rjhsha
        type: string
    type: object
info:
  contact:
    email: dmad1989@gmail.com
  description: Сервис сокращения ссылок.
  title: URLCutter API
  version: "1.0"
paths:
  /:
    post:
      consumes:
      - application/json
      operationId: deleteUserUrls
      produces:
      - plain/text
      responses:
        "202":
          description: Accepted
          schema:
            items:
              type: string
            type: array
        "400":
          description: Ошибка
          schema:
            type: string
      summary: Запрос на удаление сокращеных URL
      tags:
      - UserURLs
  /{path}:
    get:
      consumes:
      - plain/text
      operationId: redirect
      parameters:
      - description: Сокращенный url
        in: path
        name: path
        required: true
        type: string
      responses:
        "307":
          description: Переход по сокращенному URL
        "400":
          description: Ошибка
          schema:
            type: string
        "401":
          description: Ошибка авторизации
          schema:
            type: string
        "410":
          description: url was deleted
          schema:
            type: string
      summary: Переход по сокращеному URL
      tags:
      - Operate
  /api/shorten:
    post:
      consumes:
      - application/json
      operationId: cutterJSON
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/jsonobject.Response'
        "400":
          description: Ошибка
          schema:
            type: string
        "401":
          description: Ошибка авторизации
          schema:
            type: string
      summary: Запрос на сокращение URL
      tags:
      - Cut
  /api/shorten/batch:
    post:
      consumes:
      - application/json
      operationId: cutterBatch
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/jsonobject.BatchItem'
            type: array
        "400":
          description: Ошибка
          schema:
            type: string
        "401":
          description: Ошибка авторизации
          schema:
            type: string
      summary: Запрос на сокращение списка URL
      tags:
      - Cut
  /api/user/urls:
    get:
      operationId: userURLs
      produces:
      - plain/text
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jsonobject.BatchItem'
            type: array
        "204":
          description: Нет сокращенных URL
          schema:
            type: string
        "400":
          description: Ошибка
          schema:
            type: string
        "401":
          description: Ошибка авторизации
          schema:
            type: string
      summary: Все скоращенные URL текущего пользователя
      tags:
      - UserURLs
  /ping:
    get:
      consumes:
      - '*/*'
      operationId: ping
      produces:
      - plain/text
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Ошибка авторизации
          schema:
            type: string
        "500":
          description: Ошибка
          schema:
            type: string
      summary: Проверка соединения с БД
      tags:
      - Info
swagger: "2.0"
tags:
- description: '"Группа запросов для сокращения URL"'
  name: Cut
- description: '"Группа запросов для работы с URL пользователя"'
  name: UserURLs
- description: '"Группа запросов для работы с сокращенными URL"'
  name: Operate
- description: '"Группа запросов состояния сервиса"'
  name: Info